+++ docs/changes.rst
@@ -13,6 +13,7 @@ Not yet released.
 * Better logging in :wladmin:`createadmin`.
 * :ref:`addon-weblate.discovery.discovery` now reports skipped entries.
 * Adding string in a repository triggers :ref:`subscriptions`.
+* :ref:`mt-openai` better handles batch translations.
 
 **Bug fixes**